refactor: move registry builder options to builder package
