+++ openmeter/registry/builder/entitlement.go
@@ -1,21 +1,34 @@
 package registrybuilder
 
 import (
+	"log/slog"
 	"time"
 
 	"github.com/openmeterio/openmeter/openmeter/credit"
 	creditpgadapter "github.com/openmeterio/openmeter/openmeter/credit/adapter"
+	"github.com/openmeterio/openmeter/openmeter/ent/db"
 	"github.com/openmeterio/openmeter/openmeter/entitlement"
 	entitlementpgadapter "github.com/openmeterio/openmeter/openmeter/entitlement/adapter"
 	booleanentitlement "github.com/openmeterio/openmeter/openmeter/entitlement/boolean"
 	meteredentitlement "github.com/openmeterio/openmeter/openmeter/entitlement/metered"
 	staticentitlement "github.com/openmeterio/openmeter/openmeter/entitlement/static"
+	"github.com/openmeterio/openmeter/openmeter/meter"
 	"github.com/openmeterio/openmeter/openmeter/productcatalog"
 	productcatalogpgadapter "github.com/openmeterio/openmeter/openmeter/productcatalog/adapter"
 	"github.com/openmeterio/openmeter/openmeter/registry"
+	"github.com/openmeterio/openmeter/openmeter/streaming"
+	"github.com/openmeterio/openmeter/openmeter/watermill/eventbus"
 )
 
-func GetEntitlementRegistry(opts registry.EntitlementOptions) *registry.Entitlement {
+type EntitlementOptions struct {
+	DatabaseClient     *db.Client
+	StreamingConnector streaming.Connector
+	Logger             *slog.Logger
+	MeterRepository    meter.Repository
+	Publisher          eventbus.Publisher
+}
+
+func GetEntitlementRegistry(opts EntitlementOptions) *registry.Entitlement {
 	// Initialize database adapters
 	featureDBAdapter := productcatalogpgadapter.NewPostgresFeatureRepo(opts.DatabaseClient, opts.Logger)
 	entitlementDBAdapter := entitlementpgadapter.NewPostgresEntitlementRepo(opts.DatabaseClient)
